pw_software_update: Fixes majority of build config
This changes fixes some errors in the build config and adds some
Python targets as well.

+++ pw_software_update/update_bundle_accessor.cc
@@ -168,7 +168,7 @@ Status VerifyMetadataSignatures(protobuf::Bytes message,
       "Not enough number of signatures verified. Requires at least %u, "
       "verified %u",
       threshold.value(),
-      verified_count);
+      static_cast<uint32_t>(verified_count));
   return Status::Unauthenticated();
 }
